set min-width of app widget properly (#20)
closes https://github.com/dstl-lab/dsc10-tutor-jlab/issues/8

before, the tutor sidebar pane could get dragged to 0 width if you
dragged the edge of the sidebar all the way to the right. even worse,
the sidebar pane could open to 0 width, which means that a regular user
would just think that the tutor was broken.

now, the sidebar can't get resized below 250px. figuring this out took
me forever, see the comments in AppWidget.tsx for the backstory.

https://github.com/user-attachments/assets/4816d8c8-9490-42ae-ad2c-d802b345a960